requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras pyproject.toml
#
alt-profanity-check==1.7.2
    # via wills_of_words (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via httpx
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via ollama
idna==3.11
    # via
    #   anyio
    #   httpx
joblib==1.5.1
    # via
    #   alt-profanity-check
    #   scikit-learn
numpy==2.3.5
    # via
    #   scikit-learn
    #   scipy
    #   wills_of_words (pyproject.toml)
ollama==0.6.1
    # via wills_of_words (pyproject.toml)
pillow==12.0.0
    # via wills_of_words (pyproject.toml)
pydantic==2.12.5
    # via
    #   ollama
    #   wills_of_words (pyproject.toml)
pydantic-core==2.41.5
    # via pydantic
scikit-learn==1.7.2
    # via alt-profanity-check
scipy==1.16.3
    # via scikit-learn
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.15.0
    # via
    #   anyio
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
